Catch interrupted poll() in terminal console
Alternative to my own PR #8108 - catch ZMQError in run_cell, and if it's
caused by an interrupt, ignore it.

more complex, especially if we want to handle the timeout nicely as
proposed in the comments, but it's possibly also more convenient for
other users of that API. Or perhaps not - I'm not sure what makes sense
for other API consumers in this case.

Fixes gh-8105