Fill readline history before recording its length
Closes gh-49

I'm not sure why the call to refill_readline_hist() is needed here at
all - it should be called in init_readline() - but from the comments in
gh-49, the earlier call doesn't always work. It may be a race condition.
It doesn't seem to do any harm to call it again, though.