Merge old ZMQTerminalInteractiveShell class up into new prompt_toolkit code
PTInteractiveShell was inheriting from ZMQTerminalInteractiveShell,
which in turn inherited from TerminalInteractiveShell in IPython.

This merges ZTIS' code up into PIS, gets rid of ZTIS, and switches PIS
to inherit from TIS directly. PIS is renamed to ZTIS for consistency
with config.

The next step after this will be to copy code in from IPython and
decouple jupyter_console from IPython's code. I intend that to be a
separate PR.