Update execution_count before running cell
Ensure the execution count displayed next to the code is correct by
processing any IOPub messages between pressing enter and running the
code. The execution count should always be updated after an
execute_input message even if the message came from a session whose
output wouldn't be included because it is still the execution count of
this session's kernel.